requirements.txt
tqdm
numpy==1.22.3
opencv-python
pytorch==1.12.1
scipy==1.7.3
torchvision==0.13.1
timm==0.5.4